+++ testdata/de/dm/intellij/liferay/index/PortletNameIndexTest/de/dm/util/MyPortletKeys.java
@@ -1,0 +1,7 @@
+package de.dm.util;
+
+public class MyPortletKeys {
+
+    public static final String MY_PORTLET_KEY = "de_dm_util_MyPortletKey";
+
+}